refactor: reuse voice note dialog
