+++ mobile/calorie-counter/src/app/components/voice-note-dialog/voice-note-dialog.component.html
@@ -1,0 +1,64 @@
+<form class="note-form" [formGroup]="form" (ngSubmit)="send()">
+  <h2 mat-dialog-title class="header">
+    <span class="title">{{ data.title }}</span>
+  </h2>
+
+  <div mat-dialog-content class="dialog-content">
+    <mat-form-field appearance="outline" class="note-field">
+      <mat-label>{{ isHistoryClarify ? 'Текст' : 'Описание' }}</mat-label>
+      <textarea matInput
+                formControlName="note"
+                cdkTextareaAutosize
+                [attr.rows]="isHistoryClarify ? null : 4"
+                cdkAutosizeMinRows="4"
+                cdkAutosizeMaxRows="12"></textarea>
+      <button *ngIf="noteControl?.value" mat-icon-button matSuffix aria-label="Очистить" (click)="clearNote()">
+        <mat-icon>close</mat-icon>
+      </button>
+    </mat-form-field>
+
+    <mat-form-field appearance="outline" class="time-field">
+      <mat-label>{{ isHistoryClarify ? 'Время' : 'Дата и время' }}</mat-label>
+      <input matInput
+             [type]="isHistoryClarify ? 'time' : 'datetime-local'"
+             formControlName="timestamp" />
+      <mat-icon *ngIf="isHistoryClarify" matSuffix>schedule</mat-icon>
+    </mat-form-field>
+
+    <div class="mic" [class.inline]="isHistoryClarify">
+      <button *ngIf="!isHistoryClarify"
+              mat-fab color="primary"
+              matTooltip="Удерживайте для записи"
+              aria-label="Записать голос"
+              (mousedown)="startRecord($event)" (touchstart)="startRecord($event)"
+              (mouseup)="stopRecord()" (mouseleave)="stopRecord()"
+              (touchend)="stopRecord()" (touchcancel)="stopRecord()"
+              [class.recording]="!!recorder" [disabled]="transcribing">
+        <mat-icon>{{ recorder ? 'mic' : 'mic_none' }}</mat-icon>
+      </button>
+
+      <button *ngIf="isHistoryClarify"
+              mat-icon-button color="primary"
+              matTooltip="Удерживайте для записи"
+              aria-label="Записать голос"
+              (mousedown)="startRecord($event)" (touchstart)="startRecord($event)"
+              (mouseup)="stopRecord()" (mouseleave)="stopRecord()"
+              (touchend)="stopRecord()" (touchcancel)="stopRecord()"
+              [class.recording]="!!recorder" [disabled]="transcribing">
+        <mat-icon>{{ recorder ? 'mic' : 'mic_none' }}</mat-icon>
+      </button>
+
+      <div class="hint" *ngIf="!recorder">Удерживайте для записи</div>
+      <div class="hint recording" *ngIf="recorder">Запись… отпустите, чтобы остановить</div>
+    </div>
+
+    <mat-progress-bar *ngIf="transcribing" mode="indeterminate"></mat-progress-bar>
+  </div>
+
+  <div mat-dialog-actions class="actions" [class.actions--history]="isHistoryClarify">
+    <button *ngIf="isHistoryClarify" mat-button color="warn" type="button" (click)="remove()">Удалить</button>
+    <span class="spacer" *ngIf="isHistoryClarify"></span>
+    <button mat-button type="button" (click)="close()">Отмена</button>
+    <button mat-flat-button color="primary" type="submit" [disabled]="!canSend">Отправить</button>
+  </div>
+</form>
