--- mobile/calorie-counter/src/app/pages/add-meal/add-meal-note-dialog.component.html
@@ -1,38 +1,0 @@
-<h2 mat-dialog-title>Описание приёма пищи</h2>
-
-<form class="note-form" [formGroup]="form" (ngSubmit)="sendText()">
-  <div mat-dialog-content class="dialog-content">
-    <mat-form-field appearance="outline">
-      <mat-label>Описание</mat-label>
-      <textarea matInput formControlName="note" rows="4"></textarea>
-      <button *ngIf="noteControl?.value" mat-icon-button matSuffix aria-label="Очистить" (click)="clearNote()">
-        <mat-icon>close</mat-icon>
-      </button>
-    </mat-form-field>
-
-    <mat-form-field appearance="outline">
-      <mat-label>Дата и время</mat-label>
-      <input matInput type="datetime-local" formControlName="dateTime" />
-    </mat-form-field>
-
-    <div class="mic">
-      <button mat-fab color="primary"
-              matTooltip="Удерживайте для записи"
-              aria-label="Записать голос"
-              (mousedown)="startRecord($event)" (touchstart)="startRecord($event)"
-              (mouseup)="stopRecord()" (mouseleave)="stopRecord()" (touchend)="stopRecord()" (touchcancel)="stopRecord()"
-              [class.recording]="!!recorder" [disabled]="transcribing">
-        <mat-icon>{{ recorder ? 'mic' : 'mic_none' }}</mat-icon>
-      </button>
-      <div class="hint" *ngIf="!recorder">Удерживайте для записи</div>
-      <div class="hint recording" *ngIf="recorder">Запись… отпустите, чтобы остановить</div>
-    </div>
-
-    <mat-progress-bar *ngIf="transcribing" mode="indeterminate"></mat-progress-bar>
-  </div>
-
-  <div mat-dialog-actions align="end">
-    <button mat-stroked-button type="button" (click)="close()">Отмена</button>
-    <button mat-flat-button color="primary" type="submit" [disabled]="!canSend">Отправить</button>
-  </div>
-</form>
